+++ CHANGELOG.rst
@@ -53,7 +53,7 @@ Changed
   which for some reason took tens of seconds per query. (#390 @valgur)
 * Empty queries raise a ``ValueError`` immediately on the client side instead of letting the server generate it. (#390 @valgur)
 * Added stricter checks for empty keyword values in queries, which would cause server-side errors. (#390 @valgur)
-* Gracefully handle cancelled futures. (#448 @avalentino)
+* Gracefully handle cancelled futures. (#448 and #479 @avalentino)
 * Use the HTTP status instead of OData metadata to determine the online status of a product when downloading. 
   This is a workaround for the rare server-side bug of the OData info for the online status being incorrect (#467). (#469 @valgur) 
 * Set the default query timeout to 60 seconds to avoid waiting indefinitely for a response. (#475 @valgur)